refactor: move internal/camelcase to internal/strcase/camelcase
