+++ internal/strcase/camelcase/camelcase.go
